+++ packages/amplify-cli-auth/src/commands/whoami.js
@@ -11,6 +11,7 @@ export default {
 		const { initSDK } = await import('@axway/amplify-cli-utils');
 		const { sdk } = initSDK({
 			baseUrl:  argv.baseUrl,
+			clientId: argv.clientId,
 			env:      argv.env,
 			realm:    argv.realm
 		});