+++ packages/amplify-cli-utils/CHANGELOG.md
@@ -1,3 +1,8 @@
+# v4.2.0 (Dec 1, 2020)
+
+ * fix: Bumped minimum Node.js requirement to 10.19.0 to prevent warnings on install.
+ * chore: Updated dependencies.
+
 # v4.1.4 (Nov 18, 2020)
 
  * chore: Updated dependencies.